+++ src/plugins/oSnap/components/Form/TransferNFT.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
 import { shorten } from '@/helpers/utils';
 import { isAddress } from '@ethersproject/address';
+import { NFT, Network, TransferNftTransaction } from '../../types';
 import {
   createTransferNftTransaction,
   getERC721TokenTransferTransactionData,
   validateTransaction
-} from '../../index';
-import { NFT, Network, TransferNftTransaction } from '../../types';
+} from '../../utils';
 import AddressInput from '../Input/Address.vue';
 
 const props = defineProps<{
@@ -66,43 +66,43 @@ watch(selectedCollectableAddress, updateTransaction);
 </script>
 
 <template>
-    <UiSelect v-model="selectedCollectableAddress" :disabled="isReadOnly">
-      <template #label>{{ $t('safeSnap.asset') }}</template>
-      <template
-        v-if="
-          selectedCollectable &&
-          (selectedCollectable.imageUri || selectedCollectable.logoUri)
-        "
-        #image
-      >
-        <img
-          :src="selectedCollectable.imageUri || selectedCollectable.logoUri"
-          alt=""
-          class="tokenImage"
-        />
-      </template>
-      <option v-if="!collectables.length" disabled selected>
-        - {{ $t('safeSnap.noCollectibles') }} -
-      </option>
-      <option
-        v-for="(collectable, index) in collectables"
-        :key="index"
-        :value="collectable.address"
-      >
-        {{ collectable.name ?? collectable.tokenName }} #{{
-          shorten(collectable.id, 10)
-        }}
-      </option>
-    </UiSelect>
+  <UiSelect v-model="selectedCollectableAddress" :disabled="isReadOnly">
+    <template #label>{{ $t('safeSnap.asset') }}</template>
+    <template
+      v-if="
+        selectedCollectable &&
+        (selectedCollectable.imageUri || selectedCollectable.logoUri)
+      "
+      #image
+    >
+      <img
+        :src="selectedCollectable.imageUri || selectedCollectable.logoUri"
+        alt=""
+        class="tokenImage"
+      />
+    </template>
+    <option v-if="!collectables.length" disabled selected>
+      - {{ $t('safeSnap.noCollectibles') }} -
+    </option>
+    <option
+      v-for="(collectable, index) in collectables"
+      :key="index"
+      :value="collectable.address"
+    >
+      {{ collectable.name ?? collectable.tokenName }} #{{
+        shorten(collectable.id, 10)
+      }}
+    </option>
+  </UiSelect>
 
-    <AddressInput
-      v-model="recipient"
-      :disabled="isReadOnly"
-      :input-props="{
-        required: true
-      }"
-      :label="$t('safeSnap.to')"
-    />
+  <AddressInput
+    v-model="recipient"
+    :disabled="isReadOnly"
+    :input-props="{
+      required: true
+    }"
+    :label="$t('safeSnap.to')"
+  />
 </template>
 
 <style scoped>
