+++ src/plugins/oSnap/utils/index.ts
@@ -1,24 +1,7 @@
-import { isAddress } from '@ethersproject/address';
-import { JsonRpcProvider } from '@ethersproject/providers';
-import memoize from 'lodash/memoize';
-
-import getProvider from '@snapshot-labs/snapshot.js/src/utils/provider';
-
-export const mustBeEthereumAddress = memoize((address: string) => {
-  const startsWith0x = address?.startsWith('0x');
-  const isValidAddress = isAddress(address);
-  return startsWith0x && isValidAddress;
-});
-
-export const mustBeEthereumContractAddress = memoize(
-  async (network: string, address: string) => {
-    const provider = getProvider(network) as JsonRpcProvider;
-    const contractCode = await provider.getCode(address);
-
-    return (
-      contractCode && contractCode.replace(/^0x/, '').replace(/0/g, '') !== ''
-    );
-  },
-  (url, contractAddress) => `${url}_${contractAddress}`
-);
-
+export * from "./abi";
+export * from "./coins";
+export * from "./events";
+export * from "./getters";
+export * from "./proposal";
+export * from "./transactions";
+export * from "./validators";
